unit test for newAllocationRequest
