+++ code/go/0chain.net/smartcontract/storagesc/helper2_test.go
@@ -5,19 +5,37 @@ import (
 	cstate "0chain.net/chaincore/chain/state"
 	"0chain.net/chaincore/state"
 	"0chain.net/chaincore/transaction"
+	"0chain.net/core/common"
 	"0chain.net/core/datastore"
 	"0chain.net/core/encryption"
 	"0chain.net/core/util"
+	"time"
 )
 
 type mockStateContext struct {
 	ctx           cstate.StateContext
 	clientBalance state.Balance
-	//block                      *block.Block
-	store map[datastore.Key]util.Serializable
-	//sharders                   []string
-	//LastestFinalizedMagicBlock *block.Block
+	store         map[datastore.Key]util.Serializable
 }
+
+type mockBlobberYaml struct {
+	serviceCharge           float64
+	readPrice               float64
+	writePrice              float64
+	challengeCompletionTime time.Duration
+	MaxOfferDuration        time.Duration
+}
+
+var (
+	scYaml          = &scConfig{}
+	creationDate    = common.Timestamp(100)
+	approvedMinters = []string{
+		"6dba10422e368813802877a85039d3985d96760ed844092319743fb3a76712d9", // miner SC
+		"cf8d0df9bd8cc637a4ff4e792ffe3686da6220c45f0e1103baa609f3f1751ef4", // interest SC
+		"6dba10422e368813802877a85039d3985d96760ed844092319743fb3a76712d7", // storage SC
+	}
+	storageScId = approvedMinters[2]
+)
 
 func (sc *mockStateContext) SetMagicBlock(_ *block.MagicBlock)                     { return }
 func (sc *mockStateContext) GetState() util.MerklePatriciaTrieI                    { return nil }
@@ -55,7 +73,11 @@ func (sc *mockStateContext) GetBlock() *block.Block {
 func (sc *mockStateContext) SetStateContext(_ *state.State) error { return nil }
 
 func (sc *mockStateContext) GetTrieNode(key datastore.Key) (util.Serializable, error) {
-	return sc.store[key], nil
+	var val, ok = sc.store[key]
+	if !ok {
+		return nil, util.ErrValueNotPresent
+	}
+	return val, nil
 }
 
 func (sc *mockStateContext) InsertTrieNode(key datastore.Key, node util.Serializable) (datastore.Key, error) {
